Drop nineteen no-op casts that restate a field's own declared type
Each site casts a struct field to the width and signedness the field is
already declared with, in the header the file itself includes -- placement
rotation bytes (237, 287, 289, 290, 548, 594), yawByte (244),
ObjSequence.movementState (25, 646), sfxId (202), proximityRange (437),
hintTextIndex (289), startRelativePosition (75). All twelve objects
byte-identical.

Two same-shaped sites are NOT no-ops and are left in place: 590's
(s16)inner->radius and objhits's (u8)stateSrc->hitVolumeId both move bytes.

+++ src/dlls/engine/25/25.c
@@ -942,7 +942,7 @@ int dll_19_func0F(GameObject* obj, ObjSeqState* seq, char* st, void* moveHandler
         ((BaddieState*)st)->moveInputX = rest;
         ((BaddieState*)st)->moveInputZ = rest;
     }
-    if ((s8)seq->movementState != 1)
+    if (seq->movementState != 1)
     {
         seq->posOffsetX = (obj)->anim.localPosX;
         seq->posOffsetY = (obj)->anim.localPosY;
@@ -1030,7 +1030,7 @@ int dll_19_func0F(GameObject* obj, ObjSeqState* seq, char* st, void* moveHandler
         }
     }
     gDll19SeqMinDist = dist;
-    if ((s8)seq->movementState == 0)
+    if (seq->movementState == 0)
     {
         ((GroundBaddieState*)st)->subMode = 0;
         ((BaddieState*)st)->controlMode = controlMode;
